Pushin data on map

+++ app/src/main/java/com/rodrigotristany/mrjeff/data/cities/models/Timezone.kt
@@ -1,7 +1,7 @@
 package com.rodrigotristany.mrjeff.data.cities.models
 
 data class Timezone(
-    val dstOffset: Int,
-    val gmtOffset: Int,
+    val dstOffset: Float,
+    val gmtOffset: Float,
     val timeZoneId: String
 )
